Adding recipes now complete with steps, just need to make the steps editable.
Had a bug where it was unable to confirm context. Found out it was because i used set content, instead of start new activity. which obviously didnt set the context to the new activity.

The bug also did continue after that, though i suspect it was because my computer needed to be restarted, so that AS could sort itself out.